requirements.txt
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
black==25.1.0
boto3==1.38.3
botocore==1.38.3
brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
chatlas==0.7.0
ci-info==0.3.0
click==8.1.8
cohere==5.15.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
deprecated==1.2.18
distro==1.9.0
docstring-parser==0.16
duckdb==1.2.2
etelemetry==0.3.1
eval-type-backport==0.2.2
executing==2.2.0
faicons==0.2.2
faker==37.1.0
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.91.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.12.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
griffe==1.7.3
groq==0.23.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.6
isodate==0.6.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-widgets==3.0.14
kiwisolver==1.4.8
langchain==0.3.24
langchain-anthropic==0.3.12
langchain-community==0.3.22
langchain-core==0.3.56
langchain-google-genai==2.0.10
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.37
linkify-it-py==2.0.3
logfire-api==3.14.1
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
markdown2==2.5.3
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.6.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistralai==1.7.0
multidict==6.4.3
mypy-extensions==1.1.0
narwhals==1.36.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.5
nvidia-nccl-cu12==2.26.2.post1
openai==1.76.0
opentelemetry-api==1.32.1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
patsy==1.0.1
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
polars==1.28.1
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
puremagic==1.28
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-ai==0.1.6
pydantic-ai-slim==0.1.6
pydantic-core==2.33.1
pydantic-evals==0.1.6
pydantic-graph==0.1.6
pydantic-settings==2.9.1
pydot==3.0.4
pydyf==0.11.0
pygments==2.19.1
pymupdf==1.25.5
pyparsing==3.2.3
pyphen==0.17.2
pyrsm==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
pyyaml==6.0.2
pyzmq==26.4.0
questionary==2.1.0
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.7
s3transfer==0.12.0
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
setuptools==80.0.0
shapely==2.1.0
shiny==1.4.0
shinyswatch==0.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
sse-starlette==2.3.3
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
types-requests==2.32.0.20250328
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
widgetsnbextension==4.0.14
wkhtmltopdf==0.2
wrapt==1.17.2
xgboost==3.0.0
yarl==1.20.0
zipp==3.21.0
zopfli==0.2.3.post1
zstandard==0.23.0
